Ensure redirect consistency even when GitHub pages isn't
GitHub pages issues redirects to canonical urls in the form of
insecure urls (which then upgrade when requested). I'm assuming
that this is a bug and have reported it to GitHub. Specifically,
they canonicalise from `/docs/programming` to `/docs/programming/`
(which is a good thing overall) by issuing a redirect to
`http://srobo.github.io/docs/programming/` (note the `http://`
protocol). This is insecure :( and is different to the url which
our proxy is expecting, meaning that the proxy misses it.

The fix here is to detect the insecure urls in the redirects and
proxy those too.