Re-order domains to match the ordering that certbot uses
Somewhat unexpectedly, certbot seems to sort these (alphabetically?)
when it renews the certificate, such that the actual file name is
'srobo.org' rather than 'studentrobotics.org'. Happily it does
update nginx having done that, however nginx is then out of sync
with ansible. This updates ansible to be in sync again.