Pass through the full request_uri to these proxied services
It turns out that when proxy_pass is given a value with variables
in then it always proxies to exactly that value, without any of
the normal appending of the rest of the path or query. This means
we need to explicitly pass the request uri through in such cases.

Exactly what this means for processing of the content returned
from the proxied host is somewhat unclear, however testing of the
code-submitter suggests at least that its redirects work as desired.