Update sr-proxy to Ubuntu Noble 24.04
Ubuntu Focal (20.04) is nearing end of life and Mythic have offered
to upgrade it to Noble. This change updates our Vagrantfile to
match that, which I then used to check that the ansible config
would apply cleanly (which it does). Obviously this isn't quite
the same as the in-place upgrade they'll do, but is still useful.