Fix sr-competitorsvcs identity to consistently be plural
This matches how the competition services (sr-competitionsvcs)
local machine is named as well as the production name for both
machines. The mixed state that was previously here resulted in
the machine not being provisioned by ansible and being given a
mismatched hostname from what the README documented it should
have.

This appears to have been missed in 4fc2034c0082e7ebcfce606a7eae44fc3fa01d13.